requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==3.1.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
Django==4.2.2
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.1
firebase==4.0.1
Flask==2.3.2
flask-ngrok==0.0.25
frozenlist==1.3.3
future==0.18.3
gcloud==0.18.3
geographiclib==2.0
geopy==2.3.0
googleapis-common-protos==1.59.1
googlemaps==4.10.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.23.2
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
liffpy==0.0.1a3
line-bot-sdk==2.4.2
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nox==2023.4.22
npm==0.1.1
numpy==1.25.0
oauth2client==4.1.3
openai==0.27.8
openpyxl==3.1.2
optional-django==0.1.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.2.0
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
pipenv==2023.6.26
platformdirs==3.5.3
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==4.23.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.15.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyngrok==6.0.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-firebase==1.2
pytz==2023.3
pywin32==306
PyYAML==5.4.1
pyzmq==25.1.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rsa==4.9
Scrapy==2.9.0
selenium==4.10.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
texttable==1.6.7
tinycss2==1.2.1
tldextract==3.4.4
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.0
virtualenv-clone==0.5.7
w3lib==2.1.1
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.3.6
wsproto==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zope.interface==6.0
